Mere OK Clavier et Feu presque bon

+++ main.cpp
@@ -23,9 +23,10 @@
 #include <errno.h>
 //------------------------------------------------------ Include personnel
 #include "Outils.h"
+#include "Constantes.h"
 #include "GMenu.h"
 #include "Heure.h"
-#include "Constante.h"
+#include "Feu.h"
 #include "Generateur.h"
 
 ///////////////////////////////////////////////////////////////////  PRIVE
@@ -38,24 +39,9 @@
 //---------------------------------------------------- Variables statiques
 
 //------------------------------------------------------ Fonctions privées
-//static type nom ( liste de paramètres )
-// Mode d'emploi :
-//
-// Contrat :
-//
-// Algorithme :
-//
-//{
-//} //----- fin de nom
 
 //////////////////////////////////////////////////////////////////  PUBLIC
 //---------------------------------------------------- Fonctions publiques
-//type Nom ( liste de paramètres )
-// Algorithme :
-//
-//{
-//} //----- fin de Nom
-
 int main(void)
 {
 	//I_1.	Masquer les signaux (SIGINT, SIGUSR2 et SIGCHLD)
@@ -74,8 +60,8 @@ int main(void)
 	/*pid_t pidVoie_N=-1;
 	pid_t pidVoie_S=-1;
 	pid_t pidVoie_E=-1;
-	pid_t pidVoie_O-1;
-	pid_t pidFeu=-1;*/
+	pid_t pidVoie_O-1;*/
+	pid_t pidFeu=-1;
 	
 	//I_2.	Créer Boite aux lettres ArrivéeVéhicules
 	key_t cle=ftok(REFERENCE, 1);
@@ -85,16 +71,16 @@ int main(void)
 	int idSem=semget( cle, NBSEM, IPC_CREAT|DROITS);
 	
 	//Initialisation des sémaphores
-	semctl(idSem, 0, SETVAL, 1); //Semaphore 0: Semaphore associé à la mémoire partagée GestionTempo
-	semctl(idSem, 1, SETVAL, 4); //Semaphore 1: Semaphore associé à la mémoire partagée CouleurFeu (potentiellement 4 lecteurs)
+	semctl(idSem, SEMAPHORE_TEMPO, SETVAL, 1); //Semaphore 0: Semaphore associé à la mémoire partagée GestionTempo
+	semctl(idSem, SEMAPHORE_COULEUR, SETVAL, 4); //Semaphore 1: Semaphore associé à la mémoire partagée CouleurFeu (potentiellement 4 lecteurs)
 	
 	//I_4.	Créer une mémoire partagée
 	int idShm=shmget(cle, sizeof(memoirePartagee), IPC_CREAT|DROITS);
 	
 	//I_5.	Attacher la tache à la mémoire partagée
 	memoirePartagee* shm=(memoirePartagee*) shmat (idShm, NULL, 0);
-	shm->tempoEO=10;
-	shm->tempoNS=20;
+	shm->tempoEO=7;
+	shm->tempoNS=4;
 	shm->couleurAxeNS=ROUGE;
 	shm->couleurAxeEO=ROUGE;
 	
@@ -115,7 +101,7 @@ int main(void)
 	
 	if( (pidMenu=fork() )==0)
 	{ //I_8.	Créer la tâche Menu
-		GMenu();
+		GMenu(idSem, idBAL, idShm, pidGenerateur);
 	}
 	else {
 		//I_9.	Créer la tâche Heure
@@ -127,43 +113,48 @@ int main(void)
 		//I_13.	Créer une instance de tâche Voie avec le paramètre O
 
 		//I_14.	Créer la tâche Feu
+		if( (pidFeu=fork() )==0)
+		{
+			Feu(idSem, idShm);
+		}
+		else {
+			// M_1.	Attendre la réception du signal SIGCHLD de la part de Menu
+			while( waitpid(pidMenu, NULL, 0)==-1 && errno==EINTR);
 
-		// M_1.	Attendre la réception du signal SIGCHLD de la part de Menu
-		while( waitpid(pidMenu, NULL, 0)==-1 && errno==EINTR);
+			//D_1.	Envoyer signal SIGUSR2 à la tâche Heure
+			kill(pidHeure, SIGUSR2);
+			
+			//D_2.	Attendre la fin de la tache Heure
+			while( waitpid(pidHeure, NULL, 0)==-1 && errno==EINTR);
 
-		//D_1.	Envoyer signal SIGUSR2 à la tâche Heure
-		kill(pidHeure, SIGUSR2);
-		
-		//D_2.	Attendre la fin de la tache Heure
-		while( waitpid(pidHeure, NULL, 0)==-1 && errno==EINTR);
+			//D_3.	Envoyer signal SIGCONT à la tâche Générateur
+			kill(pidGenerateur, SIGCONT);
+			
+			//D_4.	Envoyer signal SIGUSR2 à la tâche Générateur
+			kill(pidGenerateur, SIGUSR2);
+			
+			//D_5.	Attendre la fin de la tache Générateur
+			while( waitpid(pidGenerateur, NULL, 0)==-1 && errno==EINTR);
+			
+			//D_6.	Envoyer signal SIGUSR2 à la tâche Feu
+			kill(pidFeu, SIGUSR2);
 
-		//D_3.	Envoyer signal SIGCONT à la tâche Générateur
-		kill(pidGenerateur, SIGCONT);
-		
-		//D_4.	Envoyer signal SIGUSR2 à la tâche Générateur
-		kill(pidGenerateur, SIGUSR2);
-		
-		//D_5.	Attendre la fin de la tache Générateur
-		while( waitpid(pidGenerateur, NULL, 0)==-1 && errno==EINTR);
-		
-		//D_6.	Envoyer signal SIGUSR2 à la tâche Feu
-		//kill(pidFeu, SIGUSR2);
+			//D_7.	Attendre la fin de la tache Feu
+			while( waitpid(pidFeu, NULL, 0)==-1 && errno==EINTR);
 
-		//D_7.	Attendre la fin de la tache Feu
-		//while( waitpid(pidFeu, NULL, 0)==-1 && errno==EINTR);
+			//D_8.	Détruire la mémoire partagée
+			shmctl(idShm, IPC_RMID, 0);
 
-		//D_8.	Détruire la mémoire partagée
-		shmctl(idShm, IPC_RMID, 0);
+			//D_9.	Détruire le tableau de sémaphores
+			semctl(idSem, 0, IPC_RMID, 0);
+			
+			//D_10.	Détruire la BAL
+			msgctl(idBAL, IPC_RMID, 0);
 
-		//D_9.	Détruire le tableau de sémaphores
-		semctl(idSem, 0, IPC_RMID, 0);
-		
-		//D_10.	Détruire la BAL
-		msgctl(idBAL, IPC_RMID, 0);
-
-		//D_11.	Autodestruction
-		TerminerApplication(true);
-		exit(0);
+			//D_11.	Autodestruction
+			TerminerApplication(true);
+			exit(0);
+		}
 	}
 	return 0;
 }
